Client   kjh

+++ Palirates/Palirates/Shaders.hlsl
@@ -227,7 +227,7 @@ PS_MULTIPLE_RENDER_TARGETS_OUTPUT PSStandard(VS_STANDARD_OUTPUT input)
 float2 PS_ObjectDepthAndID(VS_STANDARD_OUTPUT input) : SV_Target0
 {
     float viewZ = input.positionV.z;
-    float Outline_ID = (float) material_info.Outline_Color_ID;
+    float Outline_ID = material_info.Outline_Color_ID;
 
     return float2(viewZ, Outline_ID);
 }
